+++ packages/nextjs/app/page.tsx
@@ -1,97 +1,83 @@
 "use client";
-
-import type { NextPage } from "next";
 import Image from "next/image";
-import StepInstruction from "~~/components/StepInstruction/StepInstruction";
-import ChallengeCard from "~~/components/ChallengeCard/ChallengeCard";
-import { firstChallenges, lastChallenges } from "~~/data-challenges/challenges";
-import { useRouter } from "next/navigation";
-import localFont from "@next/font/local";
-import SpeedStarknetIcon from "~~/components/icons/SpeedStarknetIcon";
+import { useState } from "react";
+import { BackgroundTexture } from "~~/components/BackgroundTexture";
+import { ChallengeModal } from "~~/components/Challenges/ChallengeModal";
+import { HomeItem } from "~~/components/HomeItem";
+import { Readme } from "~~/components/Readme";
+import { RecentlyVied } from "~~/components/RecentlyVied";
+import { RoadmapModal } from "~~/components/Roadmap/RoadmapModal";
+import { VideoModal } from "~~/components/Videos/VideoModal";
+import { DATA_MENU, DATA_MENU_SOCIAL } from "~~/mockup/data";
 
-const codec = localFont({ src: "../font/codec.ttf" });
+const Home = () => {
+  const [openChallenge, setOpenChallenge] = useState(false);
+  const [openVideo, setOpenVideo] = useState(false);
+  const [openRoadmap, setOpenRoadmap] = useState(false);
+  const handleItemClick = (type: string) => {
+    if (type === "challenge") {
+      setOpenChallenge(true);
+    }
+    if (type === "video") {
+      setOpenVideo(true);
+    }
+    if (type === "roadmap") {
+      setOpenRoadmap(true);
+    }
+  };
 
-const Home: NextPage = () => {
-  const router = useRouter();
   return (
-    <div>
-      <div className="w-full flex items-center justify-center flex-col bg-landing gap-10 bg-base-100 text-primary sm:gap-3 ">
-        <div className="w-full flex flex-col items-center gap-8 sm:gap-5">
-          <div className="w-full flex items-center justify-center flex-col gap-10 sm:px-[20px] sm:gap-5 sm:text-[12px] max-w-[600px] ">
-            <Image
-              src={"/Starknet-icon.svg"}
-              alt={"icon starknet"}
-              width={50}
-              height={50}
-              className="sm:w-[30px] sm:h-[30px]"
-            />
-            <span className="text-center">
-              Learn how to build on <strong>Starknet</strong>; the superpowers
-              and the gotchas.
-            </span>
-            <h1
-              className={`text-8xl font-black text-center sm:text-6xl ${codec.className}`}
-            >
-              SPEEDRUN STARKNET
-            </h1>
-          </div>
-        </div>
-        <div className="footer-header-landing"></div>
-      </div>
-      <div className="w-full flex justify-center text-lg flex-col items-center text-primary bg-">
-        <div className="w-full px-[20px] flex justify-center flex-col items-center sm:pr-[35px]">
-          {firstChallenges.slice(0, 4).map((challenge) => (
-            <ChallengeCard
-              key={challenge.id}
-              challenge={challenge.challenge}
-              title={challenge.title}
-              description={challenge.description}
-              imageUrl={challenge.imageUrl}
-              buttonText="QUEST"
-              onButtonClick={() => router.push(`/challenge/${challenge.id}`)}
-              end={challenge.end || false}
-              border={challenge.border !== undefined ? challenge.border : true}
+    <div className="bg-[#0F0F6D] h-full relative md:px-8 px-3 md:py-6 py-3">
+      <ChallengeModal
+        isOpen={openChallenge}
+        onClose={() => setOpenChallenge(false)}
+        title="Challenges"
+      />
+      <VideoModal
+        isOpen={openVideo}
+        onClose={() => setOpenVideo(false)}
+        title="Videos"
+      />
+      <RoadmapModal
+        isOpen={openRoadmap}
+        onClose={() => setOpenRoadmap(false)}
+        title="roadmap"
+      />
+      <Image
+        src={"/homescreen/middle-screen.png"}
+        alt="logo"
+        width={634}
+        height={148}
+        className="md:w-[634px] md:h-[148px] w-[320px] h-[90px] absolute left-1/2 top-1/2 transform -translate-x-1/2 -translate-y-1/2"
+      />
+      <BackgroundTexture />
+      <div className="relative z-40 h-full">
+        <div className="flex flex-col gap-8">
+          {DATA_MENU.map((item) => (
+            <HomeItem
+              key={item.name}
+              {...item}
+              onclick={() => handleItemClick(item.type)}
             />
           ))}
         </div>
 
-        <div className=" bg-ft-join flex justify-center bg-secondary-content text-secondary sm:h-[350px]">
-          <div className="w-full px-[20px] flex justify-center">
-            <div className="max-w-[1280px] flex justify-around flex-col w-full border-l-[5px] border-base-300 sm:justify-start sm:items-center sm:border-l-[3px] lg:border-l-[3px] pt-[20px]">
-              <div className="bg-banner-join flex justify-center h-[130px] w-full text-secondary font-black text-7xl items-center sm:text-3xl sm:h-[80px]">
-                <span className="font-black font-['system-ui']">
-                  JOiN CORE-STARS
-                </span>
-              </div>
-              <div className="flex lg:justify-center">
-                <div className="max-w-[430px] w-full py-20 pl-20 sm:py-0 sm:pl-3 sm:flex lg:pl-0  lg:pt-0">
-                  <span className="sm:text-[12px] sm:text-center leading-7 sm:text-xs">
-                    The Core-Stars is a curated group of Starknet builders
-                    creating products, prototypes, and protocols to enrich the
-                    web3 ecosystem. A place to ask for support, show off your
-                    builds and meet other builders. Start crafting your Web3
-                    portfolio by submitting your DEX, onchain Game or SVG NFT
-                    build.
-                  </span>
-                </div>
-              </div>
-            </div>
+        <div className="absolute right-0 z-40 transform top-1/4 -translate-y-1/4">
+          <div className="md:block hidden">
+            <Readme />
+          </div>
+          <div className="flex flex-col mt-5 gap-8 items-end">
+            {DATA_MENU_SOCIAL.map((item) => (
+              <HomeItem
+                key={item.name}
+                {...item}
+                // onclick={() => handleItemClick(item.type)}
+              />
+            ))}
           </div>
         </div>
-        <div className="w-full px-[20px] flex justify-center flex-col items-center">
-          {lastChallenges.slice(0).map((challenge) => (
-            <ChallengeCard
-              key={challenge.id}
-              challenge={challenge.challenge}
-              title={challenge.title}
-              description={challenge.description}
-              imageUrl={challenge.imageUrl}
-              buttonText="COMING SOON"
-              onButtonClick={() => router.push(`/challenge/${challenge.id}`)}
-              end={challenge.end || false}
-              border={challenge.border !== undefined ? challenge.border : true}
-            />
-          ))}
+        <div className="md:block hidden absolute bottom-10 z-40 transform left-1/2 -translate-x-1/2 max-w-[666px] w-full">
+          <RecentlyVied onClickItem={handleItemClick} />
         </div>
       </div>
     </div>
