+++ packages/nextjs/styles/style.css
@@ -1,0 +1,10 @@
+/* #### Generated By: http://font.download #### */
+
+@font-face {
+  font-family: "Digital Numbers Regular";
+  font-style: normal;
+  font-weight: normal;
+  src:
+    local("Digital Numbers Regular"),
+    url("DigitalNumbers-Regular.woff") format("woff");
+}
